requirements.txt
streamlit==1.19.0
pandas==1.4.4
pypi==2.1
tabula==1.0.5
SQLAlchemy==1.4.46
mysql-connector==2.2.9
PyPDF2==3.0.1
Pillow==9.4.0
pdfplumber==0.8.0
docx2txt==0.8
text-unidecode==1.3
